Define CRD API for global mesh configuration
The first step in migrating our API to a CRD is to define the CRD itself. In this case, it's actually multiple CRDs.

- The top level CRD is the MeshConfigClass, which follows the GatewayClass and IngressClass pattern. This CRD simply references the controller that should be managing this class of mesh configs. This is to future proof against the case of having multiple meshes.

- The MeshConfig CRD references the aforementioned class. This CRD only contains the configuration fields that a user can change at runtime. This is the CRD that a user will edit to change settings. All fields are optional. When a user makes updates to the config, the controller watches those updates and applies them.

- The meshconfig ConfigMap contains the full mesh configuration, used for bootstrapping the controller and for informational purposes to the user. This ConfigMap is not to be managed by a user, only the controller. The controller will ensure that the full meshconfig stays in sync with what the user defines in the MeshConfig CRD.

Also fixed an issue with our helm invoker that caused caching issues, resulting in failure to create the new CRDs and use them.

The version is currently set to v1alpha2 to avoid conflicts with our existing API server. Once that's removed, we can drop it to v1alpha1.