Get lingering pod data from k8s instead of mesh
To simplify our mesh API server, we can get rid of the calls to the /resources endpoint, in favor of just using the k8s API to get the information we need. This pertains to the list of injected pods that still exist that are printed out when the mesh is removed, for informational purposes.