Add manual injection logic
In an effort to simplify our API server, we're migrating the manual injection logic into the CLI itself. This way the CLI does not need to contact our control plane at all, and can perform the injection. The control plane will import any libraries defined here for the webhook-based injection.